[REFACTOR] Removing the useless route file fro the cases cron
